# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: AdOps Intelligence Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /analytics/dashboard:
    get:
      operationId: getDashboardSummary
      tags: [analytics]
      summary: Get top-level dashboard KPIs
      parameters:
        - name: dateFrom
          in: query
          schema:
            type: ["string", "null"]
        - name: dateTo
          in: query
          schema:
            type: ["string", "null"]
      responses:
        "200":
          description: Dashboard summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardSummary"
  /analytics/profit-over-time:
    get:
      operationId: getProfitOverTime
      tags: [analytics]
      summary: Profit, revenue, and cost grouped by day
      parameters:
        - name: dateFrom
          in: query
          schema:
            type: ["string", "null"]
        - name: dateTo
          in: query
          schema:
            type: ["string", "null"]
      responses:
        "200":
          description: Time series data
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ProfitTimePoint"
  /analytics/by-client:
    get:
      operationId: getAnalyticsByClient
      tags: [analytics]
      summary: Spend, cost, profit, and margin grouped by client
      parameters:
        - name: dateFrom
          in: query
          schema:
            type: ["string", "null"]
        - name: dateTo
          in: query
          schema:
            type: ["string", "null"]
      responses:
        "200":
          description: Per-client analytics
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ClientAnalytics"
  /analytics/by-platform:
    get:
      operationId: getAnalyticsByPartner
      tags: [analytics]
      summary: Spend, cost, profit, and margin grouped by platform
      parameters:
        - name: dateFrom
          in: query
          schema:
            type: ["string", "null"]
        - name: dateTo
          in: query
          schema:
            type: ["string", "null"]
      responses:
        "200":
          description: Per-platform analytics
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PartnerAnalytics"
...
  /bills:
    get:
      operationId: listBills
      tags: [bills]
      summary: List all bills
      parameters:
        - name: clientId
          in: query
          schema:
            type: ["integer", "null"]
        - name: buyingHouseId
          in: query
          schema:
            type: ["integer", "null"]
        - name: status
          in: query
          schema:
            type: ["string", "null"]
      responses:
        "200":
          description: List of bills
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/BillSummary"
    post:
      operationId: createBill
      tags: [bills]
      summary: Create a bill
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BillInput"
      responses:
        "201":
          description: Created bill
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BillDetail"
  /bills/{id}:
    get:
      operationId: getBill
      tags: [bills]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Bill detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BillDetail"
    patch:
      operationId: updateBill
      tags: [bills]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BillInput"
      responses:
        "200":
          description: Updated bill
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BillDetail"
    delete:
      operationId: deleteBill
      tags: [bills]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
components:
  schemas:
...
    BillingRecord:
      type: object
      required: [id, platformId, buyingHouseId, costModelId, period, appsflyerPins, fraudPins, payoutRate, marginPct, salesTaxPct, remittanceTaxPct, withholdingTaxPct, createdAt]
      properties:
        id:
          type: integer
        platformId:
          type: integer
        buyingHouseId:
          type: integer
        buyingHouseName:
          type: ["string", "null"]
        clientId:
          type: ["integer", "null"]
        clientName:
          type: ["string", "null"]
        costModelId:
          type: integer
        costModelName:
          type: ["string", "null"]
        costModelPayoutRate:
          type: ["number", "null"]
        costModelMarginPct:
          type: ["number", "null"]
        period:
          type: string
        appsflyerPins:
          type: integer
        fraudPins:
          type: integer
        payoutRate:
          type: number
        marginPct:
          type: number
        forexSellingRate:
          type: ["number", "null"]
        forexBuyingRate:
          type: ["number", "null"]
        bulkDiscountPct:
          type: ["number", "null"]
        platformBulkDiscountPct:
          type: ["number", "null"]
        salesTaxPct:
          type: number
        remittanceTaxPct:
          type: number
        withholdingTaxPct:
          type: number
        createdBy:
          type: ["string", "null"]
        createdAt:
          type: string
...
    DashboardSummary:
      type: object
      required: [totalRevenue, totalCost, totalProfit, marginPct, clientCount, platformCount, campaignCount, transactionCount]
      properties:
        totalRevenue:
          type: number
        totalCost:
          type: number
        totalProfit:
          type: number
        marginPct:
          type: number
        clientCount:
          type: integer
        platformCount:
          type: integer
        campaignCount:
          type: integer
        transactionCount:
          type: integer
        revenueChange:
          type: ["number", "null"]
          description: Percent change vs previous period
        profitChange:
          type: ["number", "null"]
        costChange:
          type: ["number", "null"]
    ProfitTimePoint:
      type: object
      required: [date, revenue, cost, profit]
      properties:
        date:
          type: string
        revenue:
          type: number
        cost:
          type: number
        profit:
          type: number
    ClientAnalytics:
      type: object
      required: [clientId, clientName, revenue, cost, profit, marginPct, transactionCount]
      properties:
        clientId:
          type: integer
        clientName:
          type: string
        buyingHouse:
          type: string
        revenue:
          type: number
        cost:
          type: number
        profit:
          type: number
        marginPct:
          type: number
        transactionCount:
          type: integer
    PartnerAnalytics:
      type: object
      required: [platformId, platformName, revenue, cost, profit, marginPct, transactionCount]
      properties:
        platformId:
          type: integer
        platformName:
          type: string
        revenue:
          type: number
        cost:
          type: number
        profit:
          type: number
        marginPct:
          type: number
        transactionCount:
          type: integer
...
    BillSummary:
      type: object
      required: [id, billNumber, status, totalReceivable, totalPaid, totalPending, transactionCount, createdAt]
      properties:
        id:
          type: integer
        billNumber:
          type: string
        clientId:
          type: ["integer", "null"]
        clientName:
          type: ["string", "null"]
        buyingHouseId:
          type: ["integer", "null"]
        buyingHouseName:
          type: ["string", "null"]
        status:
          type: string
        totalReceivable:
          type: number
        totalPaid:
          type: number
        totalPending:
          type: number
        transactionCount:
          type: integer
        notes:
          type: ["string", "null"]
        createdBy:
          type: ["string", "null"]
        createdAt:
          type: string
    BillDetail:
      type: object
      required: [id, billNumber, status, totalReceivable, totalPaid, totalPending, transactions, payments, createdAt]
      properties:
        id:
          type: integer
        billNumber:
          type: string
        clientId:
          type: ["integer", "null"]
        clientName:
          type: ["string", "null"]
        buyingHouseId:
          type: ["integer", "null"]
        buyingHouseName:
          type: ["string", "null"]
        status:
          type: string
        totalReceivable:
          type: number
        totalPaid:
          type: number
        totalPending:
          type: number
        notes:
          type: ["string", "null"]
        createdBy:
          type: ["string", "null"]
        createdAt:
          type: string
        transactions:
          type: array
          items:
            $ref: "#/components/schemas/BillingRecord"
        payments:
          type: array
          items:
            type: object
            required: [paymentId, amountApplied]
            properties:
              paymentId:
                type: integer
              amountApplied:
                type: number
    BillInput:
      type: object
      required: [billingRecordIds]
      properties:
        clientId:
          type: ["integer", "null"]
        buyingHouseId:
          type: ["integer", "null"]
        billingRecordIds:
          type: array
          items:
            type: integer
        notes:
          type: ["string", "null"]
...