# lib/api-zod/src/generated/types/partnerAnalytics.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * AdOps Intelligence Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface PartnerAnalytics {
  platformId: number;
  platformName: string;
  revenue: number;
  cost: number;
  profit: number;
  marginPct: number;
  transactionCount: number;
}